restructure and update docs
